step21 complete

+++ osa4/blogilista/controllers/blogs.js
@@ -34,6 +34,18 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  //check if the deleter is owner of the blog
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({ error: 'forbidden' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
